requirements.txt
# Django & Core
Django>=4.2.0,<5.3.0
django-cors-headers>=4.0.0,<5.0.0

# Django REST Framework
djangorestframework>=3.14.0,<4.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Image processing (for ImageField)
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0
# Database adaptors (uncomment as needed)
# psycopg2-binary>=2.9.6  # PostgreSQL
# mysqlclient>=2.1.1  # MySQL

# Environment & config
python-dotenv>=1.0.0

# Security
django-storages>=1.14.0  # For production file storage (AWS, etc.)

# Utilities
requests>=2.31.0